Make `createDispatch()` also a generate function
The `Dispatch` API was designed to be able to compose multiple
generate middlewares so that they could be used as a single
function. However, the dispatch did not actually return a generate
function, you still had to wire it up as in this fakergen:

```js
module.exports.fakergen = (info) => {
  const result = dispatch.dispatch(info.method, info, info.args);
  if (result.handled) {
    return result.value;
  } else {
    return info.next();
  }
};
```

But when writing the docs, it occured to me that this was needless
ceremony.

This makes the dispatch _also_ a generate function so that you can
skip this step of manually wrapping it.
